Mid point binding on drag create
